Update gm-editor.lua
I wouldn't put this up because I didn't really do anything except remove an error message and made it open df.global by default when not targeting a unit/item. I just figure since I use it so much and it's so handy, I'd put it up here and I can see later if I can figure out whether it would be worth trying to get it merged into the main gm-editor or not.