Update reaction_other.txt
Trying to fix some of the more buggy feeling parts, using a pick to gather rocks and boulders feels right, and I tried to find a way to have the pick use a different body part but it's still not responding right. The intent being that you can turn teeth and nails into needles and picks and knives and such where you feel it makes more sense, and I think the same change will let you make a dragon skull war hammer, so, bonus!