Update plant_standard.txt
Adjusted the syndrome from the essence of fortitude to allow a dwarf to become a dorf, humans can become hurfs, and elves can become erfs, which of course means they can switch to a different race if desired. It can also be used for healing I suppose, but I couldn't figure out a simple way to have the syndrome just turn x race into y caste without adding in a bunch of the items and reactions.